Move battle related classes to own package.
