+++ app/src/main/java/noteworthyengine/MovementNode.java
@@ -1,5 +1,6 @@
 package noteworthyengine;
 
+import noteworthyengine.battle.BattleNode;
 import noteworthyframework.Coords;
 import utils.DoublePtr;
 import noteworthyframework.Node;
