+++ app/src/main/java/noteworthyengine/units/Platoon.java
@@ -4,8 +4,8 @@ import android.graphics.Color;
 
 import art.Animations;
 import art.Constants;
-import noteworthyengine.BattleNode;
-import noteworthyengine.BattleSystem;
+import noteworthyengine.battle.BattleNode;
+import noteworthyengine.battle.BattleSystem;
 import noteworthyengine.FieldNode;
 import noteworthyengine.FormationNode;
 import noteworthyengine.GridNode;
