+++ app/src/main/java/noteworthyengine/units/Zug.java
@@ -2,7 +2,7 @@ package noteworthyengine.units;
 
 import art.Animations;
 import art.Constants;
-import noteworthyengine.BattleNode;
+import noteworthyengine.battle.BattleNode;
 import noteworthyengine.FieldNode;
 import noteworthyengine.FormationNode;
 import noteworthyengine.GridNode;
